# incubator-backend/openapi.yaml
openapi: 3.0.3
info:
  title: IT Youth Talent Incubator API
  description: |
    # IT Youth Talent Incubator Backend API
    
    Welcome to the comprehensive API documentation for the IT Youth Talent Incubator platform.
    This platform connects talented IT students in Ghana with companies seeking skilled professionals.
    
    ## Features
    - **Authentication & Authorization**: Secure JWT-based authentication with role-based access control
    - **User Management**: Support for Students, Companies, and Admins with distinct profiles
    - **Job Posting & Management**: Companies can post, manage, and track jobs
    - **Application System**: Students can apply to jobs and track application status
    - **Admin Panel**: Comprehensive admin tools for user approval and platform management
    
    ## Authentication
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - General endpoints: 100 requests per 15 minutes
    - Authentication endpoints: 3-5 requests per 15 minutes
    
    ## Pagination
    List endpoints support pagination with the following query parameters:
    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 10, max: 100)
    
  version: 1.0.0
  contact:
    name: IT Youth Talent Incubator
    email: support@itfy-incubator.com
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
...
    Company:
      type: object
      properties:
        _id:
          type: string
          format: objectid
        user:
          $ref: '#/components/schemas/User'
        name:
          type: string
          example: "Tech Innovations Ghana"
        description:
          type: string
          example: "Leading software development company in West Africa"
        industry:
          type: string
          example: "Information Technology"
        website:
          type: string
          example: "https://techinnovations.com"
        social_links:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "LinkedIn"
              url:
                type: string
                example: "https://linkedin.com/company/techinnovations"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          
    # ==================
    # Job Schemas
    # ==================
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Unauthorized"
            error: "You are not authorized to perform this action"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Validation Error"
            error: "Some required fields are missing. Please fill them all."
...
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of items per page
...
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get my profile
      description: Retrieve the authenticated user's full profile
      operationId: getMyProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/Student'
                      - $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      tags:
        - Users
      summary: Update my profile
      description: Update the authenticated user's profile
      operationId: updateMyProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Profile fields to update (varies by role)
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags:
        - Users
      summary: Delete my account
      description: Soft delete the authenticated user's account
      operationId: deleteMyAccount
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /users/students:
    get:
      tags:
        - Students
      summary: List all students
      description: |
        Retrieve paginated list of student profiles.
        - Public access: Shows approved students only
        - Admin access: Shows all students with optional filters
      operationId: listStudents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StatusParam'
      responses:
        '200':
          description: Students retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/ValidationError'
  /users/students/job-seeking:
    get:
      tags:
        - Students
      summary: Get job-seeking students
      description: Retrieve students actively seeking employment
      operationId: getJobSeekingStudents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Job-seeking students retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
...
  /users/students/me/education:
    post:
      tags:
        - Students
      summary: Add education entry
      description: Add a new education entry to student profile
      operationId: addEducation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - institution
                - degree
                - field_of_study
              properties:
                institution:
                  type: string
                  example: "University of Ghana"
                degree:
                  type: string
                  example: "Bachelor of Science"
                field_of_study:
                  type: string
                  example: "Computer Science"
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                grade:
                  type: string
                  example: "3.8 GPA"
      responses:
        '201':
          description: Education entry added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Student'
  /users/students/me/education/{index}:
    put:
      tags:
        - Students
      summary: Update education entry
      description: Update an existing education entry
      operationId: updateEducation
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: index
          required: true
          schema:
            type: integer
          description: Index of the education entry to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                institution:
                  type: string
                degree:
                  type: string
                field_of_study:
                  type: string
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        '200':
          description: Education entry updated successfully
    
    delete:
      tags:
        - Students
      summary: Delete education entry
      description: Remove an education entry from profile
      operationId: deleteEducation
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: index
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Education entry deleted successfully
  /users/students/me/experience:
    post:
      tags:
        - Students
      summary: Add work experience
      description: Add a new work experience entry to student profile
      operationId: addExperience
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - company
                - position
              properties:
                company:
                  type: string
                  example: "Tech Company Ghana"
                position:
                  type: string
                  example: "Software Developer Intern"
                description:
                  type: string
                  example: "Developed web applications using React and Node.js"
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                  nullable: true
      responses:
        '201':
          description: Work experience added successfully
  /users/students/me/experience/{index}:
    put:
      tags:
        - Students
      summary: Update work experience
      description: Update an existing work experience entry
      operationId: updateExperience
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: index
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company:
                  type: string
                position:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Work experience updated successfully
    
    delete:
      tags:
        - Students
      summary: Delete work experience
      description: Remove a work experience entry from profile
      operationId: deleteExperience
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: index
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Work experience deleted successfully
  /users/students/me/cv:
    post:
      tags:
        - Students
      summary: Upload CV/Resume
      description: Upload or update student CV/Resume (PDF, DOC, DOCX - max 5MB)
      operationId: uploadCV
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - cv
              properties:
                cv:
                  type: string
                  format: binary
                  description: CV file (PDF, DOC, DOCX - max 5MB)
      responses:
        '200':
          description: CV uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      cv_url:
                        type: string
        '400':
          $ref: '#/components/responses/ValidationError'
    
    delete:
      tags:
        - Students
      summary: Delete CV/Resume
      description: Remove CV from student profile
      operationId: deleteCV
      security:
        - BearerAuth: []
      responses:
        '200':
          description: CV deleted successfully
  /users/students/me/skills:
    put:
      tags:
        - Students
      summary: Update skills
      description: Update student skills list
      operationId: updateSkills
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - skills
              properties:
                skills:
                  type: array
                  items:
                    type: string
                  example: ["JavaScript", "React", "Node.js", "MongoDB", "Python"]
      responses:
        '200':
          description: Skills updated successfully
...
  /users/students/{id}:
    get:
      tags:
        - Students
      summary: Get student by ID
      description: Retrieve detailed information about a specific student
      operationId: getStudentById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Student ID
      responses:
        '200':
          description: Student profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Student'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ==================
  # Company Endpoints
  # ==================
...